Small mods for A800 REVB (rewired SMOT to be a CPUID)
SMOT was unused. So it was rewired to provide a cheap CPUID,
which can be used by software.

As of this writing, the CPUID feature is not being used yet by
the code; we still compile changing the CPU# for each of the two chips.

In a future code revision, we should make use of this bit
so the code can determine which CPU is which at runtime,
so the same code can be burned on both chips.