More REV B fixes; lomem optimizations, STANDALONE added
Moved arrays to lower in memory to avoid bank boundaries
so indexing the array can be done bytewise. Saves on time
by switching out the 2 cycle 'movff' that loads the high byte
of the index register (FSR2H) from memory with 'clrf FSR2H'
instead, that just clears it to zero.

Also: in last commit:
    Added STANDALONE macro for debugging more easily.

    Also separated a800-config.h into separate files:

        a800-config-production.asm
	--------------------------
	This is the production version of the PIC config
	which assumes the PIC uses the external 16MHz xtal
	that is on the A800 board.

        a800-config-standalone.asm
	--------------------------
	This config will run on the PicKit4 programmer
	via the 5 connections that will both power the chip
	and program and debug it while it's running, and the
	oscilloscope can be attached to the STEP outputs
	while the debugger changes the vel[] array to change
	the velocities.